+++ Сысоев/base/TPolinom.h
@@ -28,7 +28,7 @@ public:
 	TPolinom& operator=(TPolinom& p);
 	TPolinom  operator+(TPolinom& p);
 	TPolinom  operator-(TPolinom& p);
-	int ResultInXYZ(int _x, int _y, int _z);       //результат в точке 
+	double ResultInXYZ(int _x, int _y, int _z);    //результат в точке 
 	string XYZ_Diff(char ch);                      //производная
 	string Integration(char ch);                   //интеграл
 	string GetPolin(TPolinom& p);                  //получение полинома как строки
